requirements.txt
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
db-dtypes==1.1.1
Flask==2.2.3
Flask-Cors==3.0.10
google-api-core==2.11.0
google-auth==2.16.2
google-cloud-bigquery==3.7.0
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.3
grpcio-status==1.51.3
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.3
packaging==23.0
pandas==2.0.1
proto-plus==1.22.2
protobuf==4.22.1
pyarrow==12.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
rsa==4.9
six==1.16.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
